1 speex编码，音质杂音很大，还没找到原因。

+++ 3th/speex-1.2.0/include/speex/speex_types.h
@@ -116,8 +116,8 @@
 
    typedef short spx_int16_t;
    typedef unsigned short spx_uint16_t;
-   typedef long spx_int32_t;
-   typedef unsigned long spx_uint32_t;
+   typedef int spx_int32_t;
+   typedef unsigned int spx_uint32_t;
 
 
 #elif defined(CONFIG_TI_C6X)
